(refactor) TableGroup Controller 수정
