+++ src/test/java/kitchenpos/order/acceptance/TableGroupAcceptanceTest.java
@@ -1,11 +1,14 @@
-package kitchenpos.acceptance;
+package kitchenpos.order.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import kitchenpos.AcceptanceTest;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.TableGroup;
+import kitchenpos.order.dto.OrderTableRequest;
+import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.order.dto.TableGroupRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,25 +20,28 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 public class TableGroupAcceptanceTest extends AcceptanceTest {
 
-    private OrderTable firstOrderTable;
-    private OrderTable secondOrderTable;
+    private OrderTableRequest orderTableRequest1;
+    private OrderTableRequest orderTableRequest2;
 
     @BeforeEach
     public void setUp() {
         super.setUp();
-
-        firstOrderTable = TableAcceptanceTest.주문_테이블_등록_되어있음(new OrderTable(50, true)).as(OrderTable.class);
-        secondOrderTable = TableAcceptanceTest.주문_테이블_등록_되어있음(new OrderTable(30, true)).as(OrderTable.class);
+        orderTableRequest1 = new OrderTableRequest(50, true);
+        orderTableRequest2 = new OrderTableRequest(30, true);
+        OrderTableResponse orderTableResponse1 = TableAcceptanceTest.주문_테이블_등록_되어있음(orderTableRequest1).as(OrderTableResponse.class);
+        OrderTableResponse orderTableResponse2 = TableAcceptanceTest.주문_테이블_등록_되어있음(orderTableRequest2).as(OrderTableResponse.class);
+        orderTableRequest1.setId(orderTableResponse1.getId());
+        orderTableRequest2.setId(orderTableResponse2.getId());
     }
 
     @DisplayName("단체 지정을 관리한다.")
     @Test
     void manageTableGroup() {
         // given
-        TableGroup tableGroup = new TableGroup(Arrays.asList(firstOrderTable, secondOrderTable));
+        TableGroupRequest tableGroupRequest = new TableGroupRequest(Arrays.asList(orderTableRequest1, orderTableRequest2));
 
         // when
-        ExtractableResponse<Response> createResponse = 단체_지정_생성_요청(tableGroup);
+        ExtractableResponse<Response> createResponse = 단체_지정_생성_요청(tableGroupRequest);
 
         // then
         단체_지정_생성됨(createResponse);
@@ -47,11 +53,11 @@ public class TableGroupAcceptanceTest extends AcceptanceTest {
         단체_지정_삭제됨(deleteResponse);
     }
 
-    private ExtractableResponse<Response> 단체_지정_생성_요청(TableGroup tableGroup) {
+    private ExtractableResponse<Response> 단체_지정_생성_요청(TableGroupRequest tableGroupRequest) {
         return RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(tableGroup)
+                .body(tableGroupRequest)
                 .when().post("/api/table-groups")
                 .then().log().all()
                 .extract();
